move `_save_project_config` to `FileDataContext`
